v4 changes the way the codon occupancy is calculated compared to v3:
A pause score is calculated for every peak by dividing the peak count by the ribosomal coverage of the gene that contains that peak (ribosomal occupancy of the gene is sum of peak counts on the gene / gene length).

Then, the Codon Pause Score is calculated by summing Pause scores of each Codon.

This Codon Pause Score is divided by the total number of peaks in ORFs to obtain the Normalised_codon_pause_score.

This Normalised codon pause score is then plotted vs the codon usage of B. subtilis in the same way that codon occupancy was plotted against it in v3